refactor(router): split proposals.router into sub-routers (crud, delivery, contracts)
Follows the projects.router sub-directory pattern.
Updates all frontend tRPC call sites to use new sub-router paths.
