+++ src/features/proposal-flow/dal/client/queries/use-get-finance-options.ts
@@ -3,5 +3,5 @@ import { useTRPC } from '@/trpc/helpers'
 
 export function useGetFinanceOptions() {
   const trpc = useTRPC()
-  return useQuery(trpc.proposalsRouter.getFinanceOptions.queryOptions())
+  return useQuery(trpc.proposalsRouter.crud.getFinanceOptions.queryOptions())
 }
