+++ src/features/proposal-flow/dal/client/queries/use-get-proposal.ts
@@ -6,12 +6,12 @@ import { useTRPC } from '@/trpc/helpers'
 type TRPCClient = ReturnType<typeof useTRPC>
 
 type GetProposalQueryOptions
-  = Parameters<TRPCClient['proposalsRouter']['getProposal']['queryOptions']>[1]
+  = Parameters<TRPCClient['proposalsRouter']['crud']['getProposal']['queryOptions']>[1]
 
 type RouterOutputs = inferRouterOutputs<AppRouter>
-type GetProposalOutput = RouterOutputs['proposalsRouter']['getProposal']
+type GetProposalOutput = RouterOutputs['proposalsRouter']['crud']['getProposal']
 
 export function useGetProposal(proposalId: string, token?: string, options?: GetProposalQueryOptions) {
   const trpc = useTRPC()
-  return useQuery(trpc.proposalsRouter.getProposal.queryOptions<GetProposalOutput>({ proposalId, token }, options as any))
+  return useQuery(trpc.proposalsRouter.crud.getProposal.queryOptions<GetProposalOutput>({ proposalId, token }, options as any))
 }
