+++ src/features/proposal-flow/ui/components/proposal/index.tsx
@@ -22,8 +22,8 @@ export function Proposal() {
   const proposal = useCurrentProposal()
   const { setRootEl } = useScrollRoot()
   const trpc = useTRPC()
-  const recordView = useMutation(trpc.proposalsRouter.recordView.mutationOptions())
-  const sendContract = useMutation(trpc.proposalsRouter.sendContractForSigning.mutationOptions())
+  const recordView = useMutation(trpc.proposalsRouter.delivery.recordView.mutationOptions())
+  const sendContract = useMutation(trpc.proposalsRouter.contracts.sendContractForSigning.mutationOptions())
   const ability = useAbility()
   const hasRecorded = useRef(false)
